refactor: update SizeMib of volume and filesystem type
Makes the type uint64 to handle the full scale of size. PR feedback.

+++ domain/application/state/types.go
@@ -991,11 +991,11 @@ type storageInstanceInfoForAttach struct {
 	RequestedSizeMIB           int              `db:"requested_size_mib"`
 	FilesystemOwnedMachineUUID sql.Null[string] `db:"filesystem_owned_machine_uuid"`
 	FilesystemProvisionScopeID sql.Null[int]    `db:"filesystem_provision_scope_id"`
-	FilesystemSizeMIB          sql.Null[int]    `db:"filesystem_size_mib"`
+	FilesystemSizeMIB          sql.Null[uint64] `db:"filesystem_size_mib"`
 	FilesystemUUID             sql.Null[string] `db:"filesystem_uuid"`
 	VolumeOwnedMachineUUID     sql.Null[string] `db:"volume_owned_machine_uuid"`
 	VolumeProvisionScopeID     sql.Null[int]    `db:"volume_provision_scope_id"`
-	VolumeSizeMIB              sql.Null[int]    `db:"volume_size_mib"`
+	VolumeSizeMIB              sql.Null[uint64] `db:"volume_size_mib"`
 	VolumeUUID                 sql.Null[string] `db:"volume_uuid"`
 }
